Feat: "Search 타입 input 디자인, 기능 구현"
